Create RoundingDemoRound().java
As an alternative, you can use the round() method that is supplied with Java’s Math class. The round() method returns the nearest long value. RoundingDemoRound().java shows a program that multiplies the double answer by 100, rounds it, and then divides by 100.0. The output is identical to that shown in RoundingDemo.java