chapter 13 13-6
Uses AcctNumsIn.txt & AcctNumsOut.txt



AcctNumsIn.txt has been provided and contains a list of at least 15 six-digit account numbers. Implement the ValidateCheckDigits program to read in each account number and display whether it is valid. An account number is valid only if the last digit is equal to the remainder when the sum of the first five digits is divided by 10. For example, the number 223355 is valid because the sum of the first five digits is 15, the remainder when 15 is divided by 10 is 5, and the last digit is 5. Write only valid account numbers to an output file (AcctNumsOut.txt), each on its own line.