4-11 A
Modify the class named Person that holds the following fields: two String objects for the person’s first and last name and a LocalDate object for the person’s birthdate.

Modify the class named Couple that contains two Person objects.

Modify the class named Wedding for a wedding planner that includes the date of the wedding, the Couple being married, and a String for the location.

Provide constructors for each class that accept parameters for each field, and provide get methods for each field. The TestWedding.java program has been provided for you to test the implementations of the Person, Couple, and Wedding classes.

The TestWedding.java file has been provided to test your implementation.